+++ runtime/adapters/src/lib.rs
@@ -51,9 +51,9 @@ use xcm_executor::{
 	Assets,
 };
 
-pub mod filters;
 pub mod inspect;
 pub mod xcm_exchange;
+pub mod xcm_execute_filter;
 
 #[cfg(test)]
 mod tests;
