+++ runtime/adapters/src/tests/mod.rs
@@ -1,4 +1,4 @@
 pub mod mock;
 pub mod trader;
 pub mod xcm_exchange;
-pub mod xcm_filter;
+pub mod xcm_execute_filter;
